Added a checkLegalMove function
Checks if a move made by either the opponent or player is illegal. Also made it so that this check is ran in the process_opponent_move function. The function has options to be uncommented that will provide specific prints for each error, however this messes with the referee so they are currently left commented.